pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "dflash-mlx"
version = "0.1.0"
description = "DFlash speculative decoding on Apple Silicon with MLX"
requires-python = ">=3.10"
dependencies = [
    "mlx",
    "mlx-lm",
    "huggingface-hub",
    "datasets",
    "numpy",
]

[project.scripts]
dflash-mlx = "dflash_mlx.cli:main"
dflash-mlx-bench = "dflash_mlx.benchmark_cli:main"
dflash-mlx-prepare-model = "dflash_mlx.model_prep:main"

[tool.setuptools.packages.find]
include = ["dflash_mlx*"]

[project.optional-dependencies]
charts = [
    "matplotlib",
]
video = [
    "opencv-python",
    "pillow",
]
dev = [
    "pytest",
    "matplotlib",
]
